ethical-hacking/docx-metadata-extractor/requirements.txt
python-docx